requirements.txt
Django==2.2.12
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-phone-field==1.8.0
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.1
